PRO-385: подбор профессий по Pearson по полному 6-мерному RIASEC

+++ app/schemas/admin_result.py
@@ -20,7 +20,7 @@ class AdminCareerMatch(BaseModel):
     slug: str
     name: str
     holland_code: str
-    match_score: int
+    match_score: float
     description: str
     professions: list[str]
     skills_needed: list[str]
