+++ app/schemas/student_context.py
@@ -16,7 +16,7 @@ class ContextCareer(BaseModel):
     slug: str
     name: str
     holland_code: str
-    match_score: int
+    match_score: float
     description: str = ""
     professions: list[str] = []
     skills_needed: list[str] = []
